Create common `plugin-typegpu` package for all build plugins
