+++ apps/typegpu-docs/package.json
@@ -32,7 +32,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "remeda": "^2.3.0",
-    "rollup-plugin-typegpu": "workspace:*",
+    "plugin-typegpu": "workspace:*",
     "sharp": "^0.32.5",
     "starlight-blog": "^0.12.0",
     "starlight-typedoc": "^0.17.0",
