--- packages/babel-plugin/README.md
@@ -1,16 +1,0 @@
-<div align="center">
-
-# babel-plugin-typegpu
-
-🚧 **Under Construction** 🚧 - [GitHub](https://github.com/software-mansion/TypeGPU/tree/main/packages/rollup-plugin)
-
-</div>
-
-A babel plugin for [TypeGPU](https://typegpu.com), transpiling JavaScript functions into WGSL at build-time.
-
-
-## TypeGPU is created by Software Mansion
-
-[![swm](https://logo.swmansion.com/logo?color=white&variant=desktop&width=150&tag=typegpu-github 'Software Mansion')](https://swmansion.com)
-
-Since 2012 [Software Mansion](https://swmansion.com) is a software agency with experience in building web and mobile apps. We are Core React Native Contributors and experts in dealing with all kinds of React Native issues. We can help you build your next dream product – [Hire us](https://swmansion.com/contact/projects?utm_source=typegpu&utm_medium=readme).
