+++ packages/plugin/src/index.ts
@@ -1,0 +1,9 @@
+import babelPlugin from './babel';
+import rollupPlugin from './rollup';
+
+export type { TypegpuPluginOptions } from './common';
+
+export default {
+  babel: babelPlugin,
+  rollup: rollupPlugin,
+};
