+++ packages/plugin/tsup.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'tsup';
 const { inDevMode } = initBuildScript();
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/rollup.ts', 'src/babel.ts'],
   outDir: 'dist',
   format: ['cjs', 'esm'],
   tsconfig: './tsconfig.json',
