+++ packages/rollup-plugin/package.json
@@ -50,13 +50,10 @@
     "prepare-package": "tgpu-dev-cli prepack"
   },
   "dependencies": {
-    "tinyest-for-wgsl": "workspace:~0.1.0-alpha.0",
-    "estree-walker": "^3.0.3",
-    "magic-string": "^0.30.11"
+    "plugin-typegpu": "workspace:^0.0.0"
   },
   "devDependencies": {
     "@typegpu/tgpu-dev-cli": "workspace:*",
-    "acorn": "^8.12.1",
     "rollup": "~4.12.0",
     "tsup": "^8.0.2",
     "typescript": "^5.3.3"
